requirements.txt
# Python version
python==3.8

# Web development
flask==1.1.2

# AI and machine learning
tensorflow==2.4.1
keras==2.4.3
numpy==1.19.5

# Image processing
Pillow==8.1.0

# API requests
requests==2.25.1

# XML parsing
lxml==4.6.2
<FINAL_CODE>